Update LinkedList.cpp
So Lab 2 is coming up for the 268 kiddos taking it in the Spring semester. Normally I don't go into this much work to protect my code, but this lab is especially important for them and I think they need to be able to create a Linked List on their own. Sorry friends, write your own code like the rest of us.